Split C++ and Scala files and build systems.
This gets rid of the need for custom sbt code to do all kinds of build
magic and instead uses cmake to build the C++ parts. The split requires
us to submit the javah-generated files to git and keep them up to date.
Since we don't expect to change the API much, this is an acceptable
tradeoff.